transcribe-frontend/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-on-heroku==1.1.2
django-templated-mail==1.1.1
djangorestframework==3.11.2
djoser==2.0.5
gunicorn==20.0.4
Markdown==3.3
psycopg2-binary==2.8.6
pytube==10.4.1
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
stripe==2.50.0
